created a new script for verify recommendation tests
